usable reverser for web bundle
