+++ src/constants.js
@@ -110,3 +110,29 @@ export const BUILTIN_MACROS = [
  * @type {{parse: boolean, not_found: boolean, write: boolean}}
  */
 export const DEFAULT_ALLOW = {write: false, not_found: false, parse: false}
+
+/**
+ * Default config values
+ * @type {Object}
+ */
+export const CONFIG_DEFAULTS = {
+	'lint.src': ['/'],
+	'lint.output': '/',
+	'lint.allow.not_found': false,
+	'lint.allow.write': false,
+	'lint.allow.parse': false,
+	'lint.config.indent.character': '\t',
+	'lint.config.indent.count': 1,
+	'lint.config.pre_tag_space': 1,
+	'lint.config.post_tag_space': 1,
+	'lint.config.inline_same_line': true,
+	'lint.config.keep_implicit': true,
+	'lint.config.void_inline': true,
+	'lint.config.element_preference': "preserve",
+	'lint.config.remove_empty': false,
+	'build.src': ['/'],
+	'build.output': 'html',
+	'build.allow.not_found': false,
+	'build.allow.write': false,
+	'build.allow.parse': false
+}
